requirements.txt
cairocffi==0.8.0
CairoSVG==1.0.22
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cssselect2==0.2.1
DateTime==4.2
Django==1.11.8
django-filter==1.1.0
django-rest-framework==0.1.0
djangorestframework==3.7.7
html5lib==1.0.1
idna==2.7
MySQL-python==1.2.5
-e git+https://github.com/seriyps/ru_number_to_text@3b9ccdb36bb744341ee7bec9c50b51198259c648#egg=num2t4ru
numpy==1.16.4
olefile==0.44
pdfrw==0.4
Pillow==4.3.0
pycparser==2.18
Pyphen==0.9.4
pytz==2017.3
requests==2.19.1
scikit-learn==0.20.3
scipy==1.2.2
six==1.11.0
sklearn==0.0
sorl-thumbnail==12.4.1
tinycss2==0.6.1
urllib3==1.23
WeasyPrint==0.41
webencodings==0.5.1
zope.interface==4.4.3
